cornerSquare and its constructors, modifiers, and getters
This is a concrete class, and has no sub classes inside of it as of now.
It is based on booleans in order to determine whether this cornerSquare
is a Jail, a Go, or a FreeSpace.

The only member variables are booleans: isGo, isJail, isFreeSpace

These booleans are private,  there are setters to turn a corner square
into a single one of these options at any given time...

Further,  we have getters to verify the status of each of these booleans
individually.